# InvestmentsApp.Backend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "InvestmentsApp.Backend",
    "version": "1.0"
  },
  "paths": {
    "/api/Investment": {
      "get": {
        "tags": [
          "Investment"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvestmentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvestmentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvestmentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Investment"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertInvestmentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertInvestmentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InsertInvestmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Investment/{id}": {
      "get": {
        "tags": [
          "Investment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Investment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvestmentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvestmentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvestmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Investment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Investment/byTicker/{ticker}": {
      "get": {
        "tags": [
          "Investment"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Investment/byIdTypeInvestment/{idTypeInvestment}": {
      "get": {
        "tags": [
          "Investment"
        ],
        "parameters": [
          {
            "name": "idTypeInvestment",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/TypeInvestment": {
      "get": {
        "tags": [
          "TypeInvestment"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypeInvestmentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypeInvestmentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TypeInvestmentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TypeInvestment"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertTypeInvestmentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertTypeInvestmentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InsertTypeInvestmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TypeInvestmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeInvestmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeInvestmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/TypeInvestment/{id}": {
      "get": {
        "tags": [
          "TypeInvestment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TypeInvestmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeInvestmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeInvestmentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TypeInvestment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTypeInvestmentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTypeInvestmentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTypeInvestmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TypeInvestmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeInvestmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeInvestmentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TypeInvestment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "InsertInvestmentDto": {
        "required": [
          "fechaCierre",
          "fechaCreacion",
          "fechaEntrada",
          "idTypeInvestment",
          "importeFinal",
          "importeInicial",
          "tikcker"
        ],
        "type": "object",
        "properties": {
          "tikcker": {
            "maxLength": 15,
            "minLength": 2,
            "type": "string"
          },
          "descripcion": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "importeInicial": {
            "maximum": 10000000000000000,
            "minimum": 1,
            "type": "number",
            "format": "double"
          },
          "importeFinal": {
            "maximum": 10000000000000000,
            "minimum": 1,
            "type": "number",
            "format": "double"
          },
          "fechaEntrada": {
            "type": "string",
            "format": "date-time"
          },
          "fechaCierre": {
            "type": "string",
            "format": "date-time"
          },
          "idTypeInvestment": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "InsertTypeInvestmentDto": {
        "required": [
          "descripcion",
          "fechaCreacion"
        ],
        "type": "object",
        "properties": {
          "descripcion": {
            "maxLength": 30,
            "minLength": 4,
            "type": "string"
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "InvestmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tikcker": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "importeInicial": {
            "type": "number",
            "format": "double"
          },
          "importeFinal": {
            "type": "number",
            "format": "double"
          },
          "fechaEntrada": {
            "type": "string",
            "format": "date-time"
          },
          "fechaCierre": {
            "type": "string",
            "format": "date-time"
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "idTypeInvestment": {
            "type": "integer",
            "format": "int64"
          },
          "rendimiento": {
            "type": "number",
            "format": "double"
          },
          "typeInvestmentDescripcion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TypeInvestmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UpdateInvestmentDto": {
        "required": [
          "fechaCierre",
          "fechaEntrada",
          "idTypeInvestment",
          "importeFinal",
          "importeInicial",
          "tikcker"
        ],
        "type": "object",
        "properties": {
          "tikcker": {
            "maxLength": 15,
            "minLength": 2,
            "type": "string"
          },
          "descripcion": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "importeInicial": {
            "maximum": 10000000000000000,
            "minimum": 1,
            "type": "number",
            "format": "double"
          },
          "importeFinal": {
            "maximum": 10000000000000000,
            "minimum": 1,
            "type": "number",
            "format": "double"
          },
          "fechaEntrada": {
            "type": "string",
            "format": "date-time"
          },
          "fechaCierre": {
            "type": "string",
            "format": "date-time"
          },
          "idTypeInvestment": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "UpdateTypeInvestmentDto": {
        "required": [
          "descripcion"
        ],
        "type": "object",
        "properties": {
          "descripcion": {
            "maxLength": 30,
            "minLength": 4,
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}